Added contact.js and contact.css to contact folder. Added each developers contact js file and the contacteach.css to the contactEach folder.
