+++ realtor-app/src/components/ContactEach/AmyContact.js
@@ -1,5 +1,5 @@
 import React from "react";
-import '../styles/contact-each.css';
+import './contact-each.css';
 
 import Headshot from '../images/Headshot.jpg';
 import { FaGithub } from "react-icons/fa"
