+++ realtor-app/src/components/images/Headshot.jpg
